+++ test/expect-tests/test.ml
@@ -1,10 +1,10 @@
-let%expect_test "EzSubst.string" =
+let%expect_test "EzSubst.V1.EZSUBST.string" =
   let s = "$hello$hell${hello}$\\$hello$" in
   let s =
-    Drom_lib.Ez_subst.string
+    Ez_subst.V1.EZ_SUBST.string
       ~brace:(fun _ s ->
         match s with "hello" -> "HELLO" | s -> Printf.sprintf "!{%s}" s)
-      () s
+      ~ctxt:() s
   in
   Printf.printf "%s\n%!" s;
   [%expect {| $hello$hellHELLO$$hello$ |}]